Decomposition _clean_object_storage. Create object_storage/utils.py
