+++ ch_tools/chadmin/cli/object_storage_group.py
@@ -11,7 +11,7 @@ from ch_tools.chadmin.internal.object_storage import (
     cleanup_s3_object_storage,
     s3_object_storage_iterator,
 )
-from ch_tools.chadmin.internal.system import match_ch_version
+from ch_tools.chadmin.internal.object_storage.utils import get_orphaned_objects_query, get_remote_data_paths_table, get_traverse_shadow_settings, traverse_object_storage
 from ch_tools.chadmin.internal.utils import execute_query
 from ch_tools.common import logging
 from ch_tools.common.cli.formatting import print_response
@@ -23,10 +23,6 @@ from ch_tools.common.clickhouse.config.storage_configuration import S3DiskConfig
 # And for metadata for which object is not found in S3.
 # These objects are not counted if their last modified time fall in the interval from the moment of starting analyzing.
 DEFAULT_GUARD_INTERVAL = "24h"
-# Batch size for inserts in a listing table
-# Set not very big value due to default ClickHouse 'http_max_field_value_size' settings value 128Kb
-# TODO: streaming upload in POST body while INSERT
-INSERT_BATCH_SIZE = 500
 # Use big enough timeout for stream HTTP query
 STREAM_TIMEOUT = 10 * 60
 
@@ -179,25 +175,12 @@ def _clean_object_storage(
         logging.info(
             f"Collecting objects... (Disk: '{disk_conf.name}', Endpoint '{disk_conf.endpoint_url}', Bucket: '{disk_conf.bucket_name}', Prefix: '{disk_conf.prefix}')",
         )
-        _traverse_object_storage(ctx, listing_table, from_time, to_time, prefix)
+        counter = traverse_object_storage(ctx, listing_table, from_time, to_time, prefix)
+        logging.info("Collected {} objects", counter)
 
-    remote_data_paths_table = "system.remote_data_paths"
-    if on_cluster:
-        remote_data_paths_table = (
-            f"clusterAllReplicas('{cluster_name}', {remote_data_paths_table})"
-        )
-
-    settings = ""
-    if match_ch_version(ctx, min_version="24.3"):
-        settings = "SETTINGS traverse_shadow_remote_data_paths=1"
-
-    antijoin_query = f"""
-        SELECT obj_path, obj_size FROM {listing_table} AS object_storage
-          LEFT ANTI JOIN {remote_data_paths_table} AS object_table
-          ON object_table.remote_path = object_storage.obj_path
-            AND object_table.disk_name = '{disk_conf.name}'
-        {settings}
-    """
+    remote_data_paths_table = get_remote_data_paths_table(on_cluster, cluster_name)
+    settings = get_traverse_shadow_settings(ctx)
+    antijoin_query = get_orphaned_objects_query(listing_table, remote_data_paths_table, disk_conf, settings)
     logging.info("Antijoin query: {}", antijoin_query)
 
     if dry_run:
@@ -250,53 +233,4 @@ def _print_response(ctx: Context, dry_run: bool, deleted: int, total_size: int) 
 
     print_response(
         ctx, clean_stats, default_format="table", table_formatter=_table_formatter
-    )
-
-
-def _traverse_object_storage(
-    ctx: Context,
-    listing_table: str,
-    from_time: Optional[timedelta],
-    to_time: timedelta,
-    prefix: str,
-) -> None:
-    """
-    Traverse S3 disk's bucket and put object names to the ClickHouse table.
-    """
-    obj_paths_batch: List[ObjListItem] = []
-    counter = 0
-    now = datetime.now(timezone.utc)
-
-    for obj in s3_object_storage_iterator(
-        ctx.obj["disk_configuration"], object_name_prefix=prefix
-    ):
-        if obj.last_modified > now - to_time:
-            continue
-        if from_time is not None and obj.last_modified < now - from_time:
-            continue
-
-        obj_paths_batch.append(ObjListItem(obj.key, obj.size))
-        counter += 1
-        if len(obj_paths_batch) >= INSERT_BATCH_SIZE:
-            _insert_listing_batch(ctx, obj_paths_batch, listing_table)
-            obj_paths_batch.clear()
-
-    # Insert the last batch (might be shorter)
-    if obj_paths_batch:
-        _insert_listing_batch(ctx, obj_paths_batch, listing_table)
-
-    logging.info("Collected {} objects", counter)
-
-
-def _insert_listing_batch(
-    ctx: Context, obj_paths_batch: List[ObjListItem], listing_table: str
-) -> None:
-    """
-    Insert batch of object names to the listing table.
-    """
-    batch_values = ",".join(f"('{item.path}',{item.size})" for item in obj_paths_batch)
-    execute_query(
-        ctx,
-        f"INSERT INTO {listing_table} (obj_path, obj_size) VALUES {batch_values}",
-        format_=None,
     )
